rename folder for consistency
