+++ source-code/ide-extension/src/main.ts
@@ -3,18 +3,18 @@ import { debounce } from "throttle-debounce"
 import { setState, state } from "./state.js"
 import { extractMessageCommand } from "./commands/extractMessage.js"
 import { messagePreview } from "./decorations/messagePreview.js"
-import { determineClosestPath } from "./utils/determineClosestPath.js"
+import { determineClosestPath } from "./utilities/determineClosestPath.js"
 import { setupConfig } from "@inlang/core/config"
 import { ExtractMessage } from "./actions/extractMessage.js"
-import { msg } from "./utils/message.js"
+import { msg } from "./utilities/message.js"
 import { createInlangEnv, importInlangConfig } from "./services/inlang-environment/index.js"
 import { telemetry } from "./services/telemetry/index.js"
 import { version } from "../package.json"
 import { propertiesMissingPreview } from "./decorations/propertiesMissingPreview.js"
-import { promptToReloadWindow } from "./utils/promptToReload.js"
+import { promptToReloadWindow } from "./utilities/promptToReload.js"
 import { coreUsedConfigEvent } from "@inlang/telemetry"
-import { recommendation, disableRecommendation } from "./utils/recommendation.js"
-import { createInlangConfigFile } from "./utils/createInlangConfigFile.js"
+import { recommendation, disableRecommendation } from "./utilities/recommendation.js"
+import { createInlangConfigFile } from "./utilities/createInlangConfigFile.js"
 
 export async function activate(context: vscode.ExtensionContext): Promise<void> {
 	try {
