+++ source-code/ide-extension/src/services/telemetry/implementation.ts
@@ -3,7 +3,7 @@ import { raw } from "@inlang-git/client/raw"
 import fs from "node:fs"
 import * as vscode from "vscode"
 import type { TelemetryEvents } from "./events.js"
-import { getUserId } from "../../utils/getUserId.js"
+import { getUserId } from "../../utilities/getUserId.js"
 
 export const telemetry: Omit<typeof telemetryNode, "capture"> & { capture: typeof capture } =
 	new Proxy(telemetryNode, {
