+++ source-code/ide-extension/src/utilities/createInlangConfigFile.ts
