+++ source-code/ide-extension/src/utilities/determineClosestPath.test.ts
