+++ source-code/ide-extension/src/utilities/determineClosestPath.ts
