+++ source-code/ide-extension/src/utilities/getUserId.ts
