+++ source-code/ide-extension/src/utilities/promptToReload.ts
